requirements.txt
affine==2.3.0
aiofiles==0.6.0
aiosqlite==0.16.1
attrs==20.3.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
h11==0.12.0
httptools==0.1.1
idna==3.1
iso8601==0.1.14
loguru==0.5.3
motor==2.2.0
numpy==1.20.1
passlib==1.7.4
pycparser==2.20
pydantic==1.7.3
PyJWT==2.0.1
pymongo==3.11.3
pyparsing==2.4.7
pyproj==3.0.1
python-decouple==3.3
python-dotenv==0.15.0
python-multipart==0.0.5
pytz==2020.5
PyYAML==5.4.1
rasterio==1.2.1
rasterstats==0.15.0
shapely==1.7.1
six==1.15.0
snuggs==1.4.7
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.3
uvloop==0.15.0
watchgod==0.6
websockets==8.1
geopandas==0.9.0